requirements.txt
asgiref==3.8.1
Django==5.1.4
django-js-asset==3.0.1
django-markdownx==4.0.7
gunicorn==23.0.0
Markdown==3.7
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
